Move DefaultsManager and utils into statusbar package.
