+++ packages/statusbar-extension/src/defaults/editorSyntax.tsx
@@ -10,7 +10,6 @@ import { TextItem } from '@jupyterlab/statusbar';
 import { ISignal } from '@phosphor/signaling';
 import { Token } from '@phosphor/coreutils';
 import { JupyterLabPlugin, JupyterLab } from '@jupyterlab/application';
-import { IDefaultsManager } from './manager';
 import { IEditorTracker, FileEditor } from '@jupyterlab/fileeditor';
 import { IStatusContext } from '../contexts';
 import { VDomRenderer, VDomModel } from '@jupyterlab/apputils';
@@ -21,8 +20,12 @@ import { IChangedArgs } from '@jupyterlab/coreutils';
 import { CommandRegistry } from '@phosphor/commands';
 import { JSONObject } from '@phosphor/coreutils';
 import { Menu } from '@phosphor/widgets';
-import { showPopup, Popup } from '@jupyterlab/statusbar';
-import { interactiveItem } from '@jupyterlab/statusbar';
+import {
+  IDefaultsManager,
+  interactiveItem,
+  showPopup,
+  Popup
+} from '@jupyterlab/statusbar';
 
 namespace EditorSyntaxComponent {
   export interface IProps {
