+++ packages/statusbar-extension/src/defaults/index.ts
@@ -1,5 +1,4 @@
 export * from './notebookTrust';
-export * from './manager';
 export * from './lineCol';
 export * from './fileUpload';
 export * from './kernelStatus';
