+++ packages/statusbar/src/util/set.ts
