+++ packages/statusbar/src/util/settings.ts
