+++ packages/statusbar/src/util/signal.ts
