+++ packages/statusbar/src/util/text.ts
